+++ src/main/kotlin/manage/laundry/service/dto/response/ShopServiceResponse.kt
@@ -1,12 +1,12 @@
 package manage.laundry.service.dto.response
 
-import java.math.BigDecimal
+
 
 data class ShopServiceResponse(
     val id: Int,
     val name: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Int,
     val shopId: Int
 )
 
@@ -14,5 +14,5 @@ data class ShopDetailServiceResponse(
     val id: Int,
     val name: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Int,
 )
