+++ src/main/kotlin/manage/laundry/service/entity/OrderItem.kt
@@ -1,7 +1,6 @@
 package manage.laundry.service.entity
 
 import jakarta.persistence.*
-import java.math.BigDecimal
 
 @Entity
 @Table(name = "order_items")
@@ -13,5 +12,5 @@ data class OrderItem(
     @ManyToOne @JoinColumn(name = "service_id")
     val shopService: ShopService,
     val quantity: Int,
-    val price: BigDecimal
+    val price: Int
 )
