+++ src/main/kotlin/manage/laundry/service/entity/ShopService.kt
@@ -1,7 +1,6 @@
 package manage.laundry.service.entity
 
 import jakarta.persistence.*
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 @Entity
@@ -13,7 +12,7 @@ data class ShopService(
     val shop: Shop,
     val name: String,
     val description: String,
-    val price: BigDecimal,
+    val price: Int,
     @Column(name = "created_at")
     val createdAt: LocalDateTime = LocalDateTime.now(),
     @Column(name = "updated_at")
